Move Staging before PostTyper
