+++ compiler/src/dotty/tools/dotc/transform/PostTyper.scala
@@ -167,26 +167,6 @@ class PostTyper extends MacroTransform with IdentityDenotTransformer { thisPhase
         }
     }
 
-    /** 1. If we are in an inline method but not in a nested quote, mark the inline method
-     *  as a macro.
-     *
-     *  2. If selection is a quote or splice node, record that fact in the current compilation unit.
-     */
-    private def handleMeta(sym: Symbol)(implicit ctx: Context): Unit = {
-
-      def markAsMacro(c: Context): Unit =
-        if (c.owner eq c.outer.owner) markAsMacro(c.outer)
-        else if (c.owner.isInlineMethod) {
-          c.owner.setFlag(Macro)
-        }
-        else if (!c.outer.owner.is(Package)) markAsMacro(c.outer)
-
-      if (sym.isSplice || sym.isQuote) {
-        markAsMacro(ctx)
-        ctx.compilationUnit.needsStaging = true
-      }
-    }
-
     private object dropInlines extends TreeMap {
       override def transform(tree: Tree)(implicit ctx: Context): Tree = tree match {
         case Inlined(call, _, _) =>
@@ -198,13 +178,11 @@ class PostTyper extends MacroTransform with IdentityDenotTransformer { thisPhase
     override def transform(tree: Tree)(implicit ctx: Context): Tree =
       try tree match {
         case tree: Ident if !tree.isType =>
-          handleMeta(tree.symbol)
           tree.tpe match {
             case tpe: ThisType => This(tpe.cls).withSpan(tree.span)
             case _ => tree
           }
         case tree @ Select(qual, name) =>
-          handleMeta(tree.symbol)
           if (name.isTypeName) {
             Checking.checkRealizable(qual.tpe, qual.posd)
             super.transform(tree)(ctx.addMode(Mode.Type))
